requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.8
redis==3.4.1
requests==2.22.0
urllib3==1.25.8
Flask==1.1.1
Flask-RESTful==0.3.8
python-decouple==3.1